+++ src/components/sn-ui/sn-tooltip.tsx
@@ -1,0 +1,13 @@
+import { ReactNode } from 'react'
+import { Tooltip, TooltipContent, TooltipTrigger } from '../ui/tooltip'
+
+export function SnSimpleTooltip({ trigger, content }: { trigger: ReactNode; content: string }) {
+  return (
+    <Tooltip>
+      <TooltipTrigger asChild>{trigger}</TooltipTrigger>
+      <TooltipContent>
+        <p>{content}</p>
+      </TooltipContent>
+    </Tooltip>
+  )
+}