+++ src/exports/script.index.ts
@@ -1,0 +1,6 @@
+// Scripting
+export { getAutocompleteData, buildAutocomplete } from '../utils/script-editor'
+export { useEsLint } from '../components/sn-ui/sn-script-editor/hooks/useEsLint'
+export { useInlineTern } from '../components/sn-ui/sn-script-editor/hooks/useTernInline'
+export { SnScriptEditor } from '../components/sn-ui/sn-script-editor/sn-script-editor'
+export { SnCodeMirror, type SnCodeMirrorHandle } from '../components/sn-ui/sn-script-editor/sn-code-mirror'