+++ src/exports/user.index.ts
@@ -1,0 +1,8 @@
+//User Related Components
+export { SnUserCard } from '../components/sn-user/sn-user-card'
+export { SnGroupCard } from '../components/sn-user/sn-group-card'
+export { default as SnAvatar } from '../components/sn-user/sn-avatar'
+export { SnGroupWrapper } from '../components/sn-user/sn-group-wrapper'
+
+//Types
+export type { SnUser, SnGroup } from '../types/user-schema'