+++ tsup.config.ts
@@ -1,0 +1,32 @@
+import { defineConfig } from 'tsup'
+import pkg from './package.json' assert { type: 'json' }
+
+const externals = [...Object.keys(pkg.dependencies ?? {}), ...Object.keys(pkg.peerDependencies ?? {})]
+
+export default defineConfig({
+  entry: {
+    index: 'src/index.ts',
+    amb: 'src/exports/amb.index.ts',
+    form: 'src/exports/form.index.ts',
+    table: 'src/exports/table.index.ts',
+    user: 'src/exports/user.index.ts',
+    skeleton: 'src/exports/skeleton.index.ts',
+    script: 'src/exports/script.index.ts',
+    standalone: 'src/exports/standalone.index.ts',
+  },
+  dts: true,
+  format: ['esm'],
+  outDir: 'dist',
+  clean: true,
+  splitting: true,
+  treeshake: true,
+  minify: true,
+  external: [
+    ...externals,
+    /^@codemirror\/.*/,
+    /^@replit\/codemirror-.*/,
+    /^prettier(\/.*)?$/,
+    /^eslint-linter-browserify(\/.*)?$/,
+    /^espree(\/.*)?$/,
+  ],
+})